More refactoring.
