+++ webapp/src/actions/index.js
@@ -1,6 +1,9 @@
 import { fromJS } from 'immutable';
 
-import { REQUEST_RESTAURANTS, RECEIVE_RESTAURANTS } from './actionTypes';
+import {
+    REQUEST_RESTAURANTS, RECEIVE_RESTAURANTS,
+    REQUEST_PENDING_ORDERS, RECEIVE_PENDING_ORDERS
+} from './actionTypes';
 
 const requestRestaurants = () => ({ type: REQUEST_RESTAURANTS });
 
@@ -21,5 +24,27 @@ export const fetchRestaurants = () => {
         return fetch('/api/restaurants')
             .then( response => response.json() )
             .then( json => dispatch(receiveRestaurants(json)) );
+    }
+}
+
+const requestPendingOrders = () => ({ type: REQUEST_PENDING_ORDERS });
+
+const receivePendingOrders = (json) => ({
+    type: RECEIVE_PENDING_ORDERS,
+    pendingOrders: fromJS(json.reduce(
+        (all, order) => {
+            all[order.groupId] = order;
+            return all;
+        },
+        {}
+    ))
+});
+
+export const fetchPendingOrders = () => {
+    return (dispatch) => {
+        dispatch(requestRestaurants())
+        return fetch('/api/groups/pending')
+            .then( response => response.json() )
+            .then( json => dispatch(receivePendingOrders(json)) );
     }
 }
