+++ webapp/src/components/panels/Link.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export default class Link extends React.Component {
+    constructor(props) {
+        super(props);
+        this.handleClick = this.handleClick.bind(this);
+    }
+
+    handleClick() {
+        window.open('http://' + this.props.url);
+    }
+
+    render() {
+        return (
+            <div className="link" onClick={this.handleClick}>
+                {this.props.url}
+            </div>
+        );
+    }
+}
