+++ webapp/src/components/panels/yourOrder/YourOrderPanel.jsx
@@ -1,7 +1,7 @@
-import '../../styles/panels/your-order-panel.scss';
+import '../../../styles/panels/your-order-panel.scss';
 
 import React from 'react';
-import PanelHeader from './PanelHeader';
+import PanelHeader from '../PanelHeader';
 
 export default class YourOrderPanel extends React.Component {
     render() {
