+++ webapp/src/reducers/restaurants.js
@@ -2,11 +2,13 @@ import { Map } from 'immutable';
 
 import { REQUEST_RESTAURANTS, RECEIVE_RESTAURANTS } from '../actions/actionTypes';
 
-export default (state = Map(), action) => {
+const restaurants = (state = Map(), action) => {
     switch(action.type) {
         case RECEIVE_RESTAURANTS:
             return state.merge(action.restaurants);
         default:
             return state;
     }
-}
+};
+
+export default restaurants
