+++ webapp/src/selectors/index.js
@@ -1,0 +1,3 @@
+export const getRestaurantName = (state, id) => state.restaurants.has(id.toString()) ?
+												state.restaurants.get(id.toString()).get('name')
+												: '';
